feat: implement message translation infrastructure and modular component-based history item architecture

+++ Telegram/SourceFiles/main/main_session_settings.cpp
@@ -724,9 +724,9 @@ void SessionSettings::addFromSerialized(const QByteArray &serialized) {
 						"Bad data for SessionSettings::addFromSerialized()"));
 					return;
 				}
-				ghostReadTill.emplace(
-					DeserializePeerId(peerId),
-					MsgId(int32(tillId)));
+					ghostReadTill.emplace(
+						DeserializePeerId(peerId),
+						MsgId(tillId));
 			}
 		}
 	}
